RAG project/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
einops==0.8.1
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.2
llama-index-cli==0.5.1
llama-index-core==0.14.2
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-faiss==0.5.1
llama-index-workflows==2.2.2
llama-parse==0.6.54
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.3
openai==1.107.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
